requirements.txt
flask
openai-whisper
torch
numpy
ffmpeg-python
python-dotenv
Werkzeug 